4.21*: Blog List Expansion, step 9

+++ Part 4/controllers/blogs.js
@@ -3,13 +3,6 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if(authorization && authorization.startsWith('Bearer ')){
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
 
 blogsRouter.get('/',async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name : 1, id : 1 })
@@ -54,12 +47,26 @@ blogsRouter.post('/',async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request,response,next) => {
-  try {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
-  } catch (error){
-    next(error)
+  
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if(!decodedToken.id){
+    return response.status(401).json({ error: 'token missing or invalid '})
   }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if(!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if(blog.user.toString() !== decodedToken.id.toString()){
+    return response.status(401).json({ error: 'only the creator can delete this blog' })
+  }
+
+  await Blog.findByIdAndDelete(request.params.id)
+  response.status(204).end()
+ 
 })
 
 blogsRouter.put('/:id', async (request,response,next) => {
